+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+### [4.2.4](https://github.com/SelectQuoteLabs/SQForm/compare/v4.2.3...v4.2.4) (2021-03-22)
+
+
+### Bug Fixes
+
+* 🐛 run date validation on date picker window close ([4088b30](https://github.com/SelectQuoteLabs/SQForm/commit/4088b30982b7b74dfbb3edd809d576f4900149eb)), closes [#127](https://github.com/SelectQuoteLabs/SQForm/issues/127)
+
+### [4.2.3](https://github.com/SelectQuoteLabs/SQForm/compare/v4.2.2...v4.2.3) (2021-03-22)
+
+
+### Bug Fixes
+
+* 🐛 DialogContent overflow causing double scrollbars ([e99f010](https://github.com/SelectQuoteLabs/SQForm/commit/e99f01054f6f7c1dd6691e911bf28f92259548fb)), closes [#139](https://github.com/SelectQuoteLabs/SQForm/issues/139)
+
+### [4.2.2](https://github.com/SelectQuoteLabs/SQForm/compare/v4.2.1...v4.2.2) (2021-03-22)
+
+
+### Bug Fixes
+
+* 🐛 setFieldValue now updates inputValue on SQFormAutocomple ([c6c4a4e](https://github.com/SelectQuoteLabs/SQForm/commit/c6c4a4ed458bd8f728db27ff3f660b1ab6f82bab)), closes [#144](https://github.com/SelectQuoteLabs/SQForm/issues/144)
+
 ### [4.2.1](https://github.com/SelectQuoteLabs/SQForm/compare/v4.2.0...v4.2.1) (2021-03-18)